Fix Sonar Bad Smells

+++ backend/src/main/java/com/mlb/mlbportal/services/player/PlayerServiceOperations.java
@@ -112,7 +112,7 @@ public class PlayerServiceOperations {
             pitcher.setEra(0.0);
             return;
         }
-        double era = (double) (pitcher.getRunsAllowed() * 9) / innings;
+        double era = (pitcher.getRunsAllowed() * 9) / innings;
         pitcher.setEra(truncateToTwoDecimals(era));
     }
 
@@ -122,7 +122,7 @@ public class PlayerServiceOperations {
             pitcher.setWhip(0.0);
             return;
         }
-        double whip = (double) (pitcher.getWalks() + pitcher.getHitsAllowed()) / innings;
+        double whip = (pitcher.getWalks() + pitcher.getHitsAllowed()) / innings;
         pitcher.setWhip(truncateToThreeDecimals(whip));
     }
